--- cmd/netcalc/main.go
@@ -1,8 +1,0 @@
-package main
-
-func main() {
-	err := Execute()
-	if err != nil {
-		return
-	}
-}
